+++ internal/gateway/types.go
@@ -93,7 +93,7 @@ type createLoanRequestRequest struct {
 	AccountNumber    string `json:"account_number" binding:"required"`
 	LoanType         string `json:"loan_type" binding:"required"`
 	Amount           int64  `json:"amount" binding:"required"`
-	RepaymentPeriod  int32  `json:"repayment_period" binding:"required"`
+	RepaymentPeriod  int64  `json:"repayment_period" binding:"required"`
 	Currency         string `json:"currency" binding:"required"`
 	Purpose          string `json:"purpose"`
 	Salary           int64  `json:"salary"`
